backend/requirements.txt
Django>=4.2.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.2
graphene-django>=3.0.0
celery>=5.3.0
redis>=4.5.0
python-decouple>=3.8
Pillow>=9.5.0
